requirements.txt
# Python version
python>=3.8,<3.12

# Project dependencies
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pip==22.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
setuptools==59.6.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1

